Add ResetPasswordEmailLink
