+++ src/login/components/LoginApp/ResetPassword/ResetPasswordMain.js
@@ -1,71 +1,27 @@
 import React, { useEffect } from "react";
-import { withRouter } from "react-router-dom";
+import { withRouter, Route  } from "react-router-dom";
 import PropTypes from "prop-types";
 import i18n from "../../../translation/InjectIntl_HOC_factory";
-import { useSelector, useDispatch, connect } from 'react-redux';
-import { getResetPassword, postEmailLink } from "../../../redux/actions/resetPasswordActions"
-import { Field, reduxForm } from "redux-form";
-import Form from "reactstrap/lib/Form";
-import CustomInput from "../../Inputs/CustomInput";
-import EduIDButton from "../../../../components/EduIDButton";
-import { validate } from "../../../app_utils/validation/validateEmail";
+import { useSelector, useDispatch } from 'react-redux';
+import { getResetPassword } from "../../../redux/actions/resetPasswordActions"
+import ResetPasswordEmailLink from "./ResetPasswordEmailLink";
+import { createBrowserHistory } from "history";
 
-let EmailForm = (props) => (
-  <Form id="reset-password-form" role="form" onSubmit={props.sendLink}>
-    <Field
-      type="email"
-      name="email"
-      label={props.translate("signup.registering-input")}
-      componentClass="input"
-      id="email-input"
-      component={CustomInput}
-      translate={props.translate}
-      placeholder="example@email.com"
-    />
-    <EduIDButton
-      className="settings-button"
-      id="reset-password-button"
-      disabled={props.invalid}
-      onClick={props.sendLink}
-    >
-      {props.translate("resetpw.send-link")}
-    </EduIDButton>
-  </Form>
-);
-
-EmailForm = reduxForm({
-  form: "reset-pass-email-form",
-  validate,
-})(EmailForm);
-
-EmailForm = connect(() => ({
-  enableReinitialize: true,
-  destroyOnUnmount: false,
-}))(EmailForm);
+export const history = createBrowserHistory();
 
 const ResetPasswordMain = (props) => {
+  console.log("[props]",props)
   const dispatch = useDispatch();
-  const csrf_token = useSelector(state => state.resetPassword.csrf_token);
+  const csrf_token = useSelector(state => state.config.csrf_token);
 
-  const sendLink = () => {
-    const email = document.querySelector("input[name='email']").value;
-    dispatch(postEmailLink(email));
-  };
-  
   useEffect(() => {
     dispatch(getResetPassword());
-  }, [csrf_token])
+  }, [csrf_token]);
 
   return (
-    <>
-      <p className="heading">{props.translate("resetpw.heading-add-email")}</p>
-      <EmailForm sendLink={sendLink} {...props} />
-      <div className="return-login-link">
-        <a href={"/login"}>
-          {props.translate("resetpw.return-login")}
-        </a>
-      </div>
-    </>
+    <div>
+      <Route path={`/reset-password/`} component={ResetPasswordEmailLink} />
+    </div>
   );
 }
